+++ historyki_roblox/gpt_relayer.py
@@ -28,7 +28,7 @@ class GtpRelayer:
 
     def generate_image(self, prompt: str, n: int=1, size: int=256, response_format: str='url') -> List[str]:
         try:
-            response = openai.Image.create(prompt=prompt, n=n, size=f'{size}x{size}')
+            response = openai.Image.create(prompt=prompt, n=n, size=f'{size}x{size}', response_format=response_format)
             image = response['data'][0][response_format]
         except Exception as ex:
             raise GtpRelayerException(
